:sparkles: Added Version option in input file
The csv admit a second parameter called "Version", the log is changed and divided in "whitelist with compliant version", "white list with uncompliant version", "not  in whitelist"
The script is still compliant with files without the second parameter for retrocompatibility. (despite the new log).

Fixed a bug where the notwhitelist softwares logged all the software ignoring the whitelist

If the software has not the version put "ND" (not defined" as version parameter